make pep8 updates
